space hndling v0.1

+++ src/main/java/de/mpg/imeji/logic/controller/SpaceController.java
@@ -1,22 +1,16 @@
 package de.mpg.imeji.logic.controller;
 
+import com.google.common.base.Function;
+import com.google.common.collect.Lists;
 import de.mpg.imeji.exceptions.BadRequestException;
 import de.mpg.imeji.exceptions.ImejiException;
-import de.mpg.imeji.exceptions.UnprocessableError;
 import de.mpg.imeji.logic.Imeji;
-import de.mpg.imeji.logic.auth.authorization.AuthorizationPredefinedRoles;
+import de.mpg.imeji.logic.ImejiSPARQL;
 import de.mpg.imeji.logic.reader.ReaderFacade;
-import de.mpg.imeji.logic.search.Search;
-import de.mpg.imeji.logic.search.SearchFactory;
-import de.mpg.imeji.logic.search.SearchResult;
-import de.mpg.imeji.logic.search.query.URLQueryTransformer;
-import de.mpg.imeji.logic.search.vo.SearchQuery;
-import de.mpg.imeji.logic.search.vo.SortCriterion;
+import de.mpg.imeji.logic.search.query.SPARQLQueries;
 import de.mpg.imeji.logic.util.ObjectHelper;
 import de.mpg.imeji.logic.util.StringHelper;
-import de.mpg.imeji.logic.vo.Item;
-import de.mpg.imeji.logic.vo.Space;
-import de.mpg.imeji.logic.vo.User;
+import de.mpg.imeji.logic.vo.*;
 import de.mpg.imeji.logic.writer.WriterFacade;
 import de.mpg.imeji.presentation.util.PropertyReader;
 import org.apache.commons.io.FileUtils;
@@ -27,7 +21,6 @@ import java.io.File;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.net.URI;
-import java.util.ArrayList;
 import java.util.List;
 
 import static com.google.common.io.Files.copy;
@@ -44,11 +37,8 @@ public class SpaceController extends ImejiController {
     private static final Logger LOGGER = LoggerFactory
             .getLogger(SpaceController.class);
 
-
-    private static final ReaderFacade reader = new ReaderFacade(
-            Imeji.spaceModel);
-    private static final WriterFacade writer = new WriterFacade(
-            Imeji.spaceModel);
+    private static final ReaderFacade reader = new ReaderFacade(Imeji.spaceModel);
+    private static final WriterFacade writer = new WriterFacade(Imeji.spaceModel);
 
     /**
      * The directory path where files are stored
@@ -69,7 +59,7 @@ public class SpaceController extends ImejiController {
                     .getAbsolutePath());
             storageUrl = StringHelper.normalizeURI(PropertyReader
                     .getProperty("imeji.instance.url"))
-                    + "file"
+                    + "spaces/file"
                     + StringHelper.urlSeparator;
         } catch (Exception e) {
             throw new RuntimeException(
@@ -85,21 +75,23 @@ public class SpaceController extends ImejiController {
      */
     public URI create(Space space, User user)
             throws ImejiException {
-        return createAskValidate(space, user, true);
+        return create(space, user, true);
     }
     public URI createNoValidate(Space space, User user)
             throws ImejiException {
-        return createAskValidate(space, user, false);
+        return create(space, user, false);
     }
 
-    public URI createAskValidate(Space space, User user, boolean validate) throws ImejiException {
+    public URI create(Space space, User user, boolean validate) throws ImejiException {
         if(validate){
             validate(space, user);
         }
+        space.setStatus(Properties.Status.RELEASED);
         writeCreateProperties(space, user);
-        GrantController gc = new GrantController();
+        //TODO: here is future grants definitions
+        /*GrantController gc = new GrantController();
         gc.addGrants(user, AuthorizationPredefinedRoles.admin(space.getId()
-                .toString(), null), user);
+                .toString(), null), user);*/
         writer.create(WriterFacade.toList(space), user);
         return space.getId();
     }
@@ -147,11 +139,12 @@ public class SpaceController extends ImejiController {
     private String update(File toCopy, String path) throws IOException {
 
         File f = new File(path);
-        if (f.exists()) {
-            removeFile(f);
+        if (f.getParentFile().exists()) {
+            //clean space dir
+            FileUtils.cleanDirectory(f.getParentFile());
         } else {
+            //create space dir
             f.getParentFile().mkdirs();
-            f.createNewFile();
         }
         copy(toCopy, f);
         return f.getAbsolutePath();
@@ -169,13 +162,13 @@ public class SpaceController extends ImejiController {
      */
     public String generateUrl(String id, String filename) {
         filename = StringHelper.normalizeFilename(filename);
-        return storageUrl + id + StringHelper.urlSeparator
+        return storageUrl + id
                 + StringHelper.urlSeparator
                 + filename;
     }
 
     /**
-     * Transform and url to a file system path
+     * Transform an url to a file system path
      *
      * @param url
      * @return
@@ -183,7 +176,6 @@ public class SpaceController extends ImejiController {
     public String transformUrlToPath(String url) {
         return URI.create(url).getPath().replace(URI.create(storageUrl).getPath(), storagePath).replace(
                 StringHelper.urlSeparator, StringHelper.fileSeparator);
-        // .replace(filename, StringHelper.normalizeFilename(filename));
     }
 
     /**
@@ -198,35 +190,26 @@ public class SpaceController extends ImejiController {
     }
 
 
-    /**
-     * Remove a single {@link File}
-     *
-     * @param f
-     */
-    public void removeFile(File f) {
-        if (f.exists()) {
-            boolean deleted = FileUtils.deleteQuietly(f);
-            if (!deleted) {
-                throw new RuntimeException(
-                        "Impossible to delete the existing file.");
-            }
-        }
-    }
 
     /**
-     * Remove a single {@link File}
+     * Remove space file storage
      *
-     * @param uri
+     * @param space
      * @throws IOException
      */
-    public void removeDirectory(URI uri) throws IOException {
-        File f = new File(transformUrlToPath(uri.toURL().toString())).getParentFile();
+    public void removeFile(Space space) throws IOException {
+        if (space == null || space.getLogoUrl() == null)
+            return;
+        String url = space.getLogoUrl().toURL().toString();
+        if (isNullOrEmptyTrim(url))
+            return;
+        File f = new File(transformUrlToPath(url)).getParentFile();
         if (f.exists())
             FileUtils.deleteDirectory(f);
     }
 
     public void validate (Space space, User u) throws ImejiException {
-        if ( isNullOrEmptyTrim(space.getTitle())) {
+        if (isNullOrEmptyTrim(space.getTitle())) {
             throw new BadRequestException("error_space_need_title");
         }
     }
@@ -257,27 +240,28 @@ public class SpaceController extends ImejiController {
                 new Space());
     }
 
+
     /**
-     * Retrieve spaces filtered by query
+     * Retrieve all imeji {@link Album}
      *
-     * @param user
-     * @param q
      * @return
      * @throws ImejiException
      */
-    public List<Space> retrieve(User user, String q) throws ImejiException {
-        List<Space> aList = new ArrayList<>();
-        try {
-            for (String spaceId: search(!isNullOrEmptyTrim(q) ? URLQueryTransformer.parseStringQuery(q) : null, user, null, 0, 0).getResults()) {
-                aList.add(retrieve(URI.create(spaceId), user));
-            }
-        } catch (Exception e) {
-            throw new UnprocessableError("Cannot retrieve spaces:", e);
-
-        }
-        return aList;
+    public List<Space> retrieveAll() throws ImejiException {
+        return Lists.transform(ImejiSPARQL.exec(SPARQLQueries.selectSpaceAll(),
+                        Imeji.spaceModel),
+                new Function<String, Space>() {
+                    @Override
+                    public Space apply(String id) {
+                        try {
+                            return retrieve(URI.create(id), Imeji.adminUser);
+                        } catch (ImejiException e) {
+                            LOGGER.info("Cannot retrieve space: " + id );
+                        }
+                        return null;
+                    }
+                });
     }
-
 
     /**
      * Retrieve an {@link Space} without its {@link Item}
@@ -299,23 +283,8 @@ public class SpaceController extends ImejiController {
      * @throws ImejiException
      */
     public void delete(Space space, User user) throws ImejiException, IOException {
-        removeDirectory(space.getLogoUrl());
+        removeFile(space);
         writer.delete(WriterFacade.toList(space), user);
     }
 
-    /**
-     * Search for space
-     *
-     * @param searchQuery
-     * @param user
-     * @param sortCri
-     * @param limit
-     * @param offset
-     * @return
-     */
-    public SearchResult search(SearchQuery searchQuery, User user,
-                               SortCriterion sortCri, int limit, int offset) {
-        Search search = SearchFactory.create(Search.SearchType.SPACE);
-        return search.search(searchQuery, sortCri, user);
-    }
 }
